Refactor game.js into three logical sub-scripts
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/scimbe/ExampleLayerInMotion?shareId=XXXX-XXXX-XXXX-XXXX).
